% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrapCI.R
\name{boostrapCI}
\alias{boostrapCI}
\title{Procedure to obtain bootstrap CI of RMPW estimator in causal decomposition
problems}
\usage{
boostrapCI(
  G,
  Y,
  gamma = 0,
  w,
  alpha = 0.05,
  estimand = "point",
  parallel = TRUE,
  B = 1000
)
}
\arguments{
\item{G}{Indicator of unmodifiable group or characteristic}

\item{Y}{Outcome}

\item{gamma}{Sensitivity parameter (log of MSM Lambda)}

\item{w}{fitted RMPW weights}

\item{alpha}{Significance level}

\item{estimand}{Target estimand, either "point", "reduction" or "residual"}

\item{parallel}{Logical indicating whether to use parallel processing}

\item{B}{Number of bootstrap samples}
}
\value{
Extrema (an interval) of the RMPW estimator.
}
\description{
Procedure to obtain bootstrap CI of RMPW estimator in causal decomposition
problems
}
