% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompAmplify.R
\name{decompAmplify}
\alias{decompAmplify}
\title{Function to perform amplification of sensitivity parameter}
\usage{
decompAmplify(
  G,
  Z,
  XA,
  XN,
  Y,
  mu_10,
  Lambda,
  e1,
  e0,
  loco_weights,
  trim = 0.01,
  allowable = TRUE,
  stab = TRUE
)
}
\arguments{
\item{G}{Indicator of unmodifiable group or characteristic}

\item{Z}{Treatment}

\item{XA}{Allowable covariates WITHOUT INTERCEPT}

\item{XN}{Non-allowable covariates WITHOUT INTERCEPT}

\item{Y}{Outcome}

\item{mu_10}{Point estimate of the target estimand}

\item{Lambda}{Sensitivity parameter (The MSM Lambda)}

\item{e1}{Propensity score for G = 1}

\item{e0}{Propensity score for G = 0}

\item{loco_weights}{Leave-one-covariate-out imbalance terms for each covariate}

\item{trim}{Trimming proportion}

\item{allowable}{Logical indicating whether to use allowability framework}

\item{w}{RMPW weights}
}
\value{
\code{data.frame} with each covariate's imbalance and beta_u term: informal calibration
}
\description{
Function to perform amplification of sensitivity parameter
}
