% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRMPW.R
\name{estimateRMPW}
\alias{estimateRMPW}
\title{Helper function to estimate RMPW weights}
\usage{
estimateRMPW(G, Z, Y, XA, XN, trim0 = 0.05, trim1 = 0.05, allowable = TRUE)
}
\arguments{
\item{G}{Indicator of unmodifiable group or characteristic}

\item{Z}{Treatment}

\item{Y}{Outcome}

\item{XA}{Allowable covariates WITHOUT INTERCEPT}

\item{XN}{Non-allowable covariates WITHOUT INTERCEPT}

\item{trim0}{Trimming proportion for e_0}

\item{trim1}{Trimming proportion for e_1}

\item{allowable}{Logical indicating whether to use allowability framework}
}
\value{
w_rmpw RMPW weights, e_g group propensity scores
}
\description{
Helper function to estimate RMPW weights
}
