% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtrema.R
\name{getExtrema}
\alias{getExtrema}
\title{Fast solution to obtain extrema of RMPW estimator in causal decomposition
problems}
\usage{
getExtrema(G, Y, gamma = 0, w, estimand = "point", stab = TRUE, RD = TRUE)
}
\arguments{
\item{G}{Indicator of unmodifiable group or characteristic}

\item{Y}{Outcome}

\item{gamma}{Sensitivity parameter (log of MSM Lambda)}

\item{w}{fitted RMPW weights}

\item{estimand}{Target estimand, either "point", "reduction" or "residual"}

\item{stab}{Logical indicating whether to use stabilized weights}

\item{RD}{logical indicating whether to use risk diff or risk ratio (RR)}
}
\value{
Extrema (an interval) of the RMPW estimator.
}
\description{
Fast solution to obtain extrema of RMPW estimator in causal decomposition
problems
}
