% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informalAmplify.R
\name{informalAmplify}
\alias{informalAmplify}
\title{Function to perform amplification of sensitivity parameter}
\usage{
informalAmplify(
  G,
  Z,
  XA,
  XN,
  Y,
  Lambda,
  trim = 0.05,
  allowable = FALSE,
  stab = TRUE
)
}
\arguments{
\item{G}{Indicator of unmodifiable group or characteristic}

\item{Z}{Treatment}

\item{XA}{Allowable covariates WITHOUT INTERCEPT}

\item{XN}{Non-allowable covariates WITHOUT INTERCEPT}

\item{Y}{Outcome}

\item{Lambda}{Sensitivity parameter (The MSM Lambda)}

\item{trim}{Trimming proportion}

\item{allowable}{Logical indicating whether to use allowability framework}
}
\value{
\code{data.frame} with each covariate's imbalance and beta_u term: informal calibration
}
\description{
Function to perform amplification of sensitivity parameter
}
