% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrapCI.R
\name{boostrapCI}
\alias{boostrapCI}
\title{Procedure to obtain bootstrap CI of RMPW estimator in causal decomposition
problems}
\usage{
boostrapCI(
  G,
  Z,
  Y,
  XA,
  XN,
  gamma = 0,
  w,
  alpha = 0.05,
  estimand = "point",
  parallel = TRUE,
  B = 1000,
  stab = TRUE,
  allowable = FALSE,
  trim = 0.05
)
}
\arguments{
\item{G}{Indicator of unmodifiable group or characteristic}

\item{Z}{Treatment}

\item{Y}{Outcome}

\item{XA}{Allowable covariates}

\item{XN}{Non-allowable covariates}

\item{gamma}{Sensitivity parameter (log of MSM Lambda)}

\item{w}{fitted RMPW weights}

\item{alpha}{Significance level}

\item{estimand}{Target estimand, either "point", "reduction" or "residual"}

\item{parallel}{Logical indicating whether to use parallel processing}

\item{B}{Number of bootstrap samples}

\item{allowable}{Logical indicating whether to use allowability framework}

\item{trim}{Trimming proportion}
}
\value{
Bootstrap CI of the desired estimator.
}
\description{
Procedure to obtain bootstrap CI of RMPW estimator in causal decomposition
problems
}
