% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRMPW.R
\name{estimateRMPW}
\alias{estimateRMPW}
\title{Helper function to estimate RMPW weights}
\usage{
estimateRMPW(G, Z, Y, XA, XN, trim = 0.01, allowable = TRUE)
}
\arguments{
\item{G}{Indicator of unmodifiable group or characteristic}

\item{Z}{Treatment}

\item{Y}{Outcome}

\item{XA}{Allowable covariates WITHOUT INTERCEPT}

\item{XN}{Non-allowable covariates WITHOUT INTERCEPT}

\item{trim}{Trimming proportion}

\item{allowable}{Logical indicating whether to use allowability framework}
}
\value{
RMPW weights
}
\description{
Helper function to estimate RMPW weights
}
