% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBIasBounds.R
\name{getBiasBounds}
\alias{getBiasBounds}
\title{Function to obtain bias bounds for amplification}
\usage{
getBiasBounds(
  G,
  Z,
  XA,
  XN,
  Y,
  w,
  Lambda,
  trim = 0.05,
  allowable = FALSE,
  stab = TRUE
)
}
\description{
Function to obtain bias bounds for amplification
}
