% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiasBounds.R
\name{getBiasBounds}
\alias{getBiasBounds}
\title{Function to obtain bias bounds for amplification}
\usage{
getBiasBounds(
  G,
  Z,
  XA,
  XN,
  Y,
  w,
  Lambda,
  trim = 0.05,
  allowable = FALSE,
  stab = TRUE
)
}
\arguments{
\item{G}{Indicator of unmodifiable group or characteristic}

\item{Z}{Treatment}

\item{XA}{Allowable covariates}

\item{XN}{Non-allowable covariates}

\item{Y}{Outcome}

\item{w}{RMPW weights}

\item{Lambda}{Sensitivity parameter (The MSM Lambda)}

\item{trim}{Trimming proportion}

\item{allowable}{Logical indicating whether to use allowability framework}
}
\value{
A list of 3 items: the lower and upper bounds: mu_10(h) - mu_10,
the point estimate mu_10, the RMPW weights
}
\description{
Function to obtain bias bounds for amplification
}
